Added MyPDO Class
MyPDO class is a wrapper which contains certain functions which are 1. a constructor 2. run
We are using a constructor to connect to the database
We are using run to execute the query
We are explicitly creating two different methods except for native methods given in php because it makes our job easy to connect and execute the query